+++ Data/Data.lua
@@ -213,7 +213,7 @@ end
 --         elseif string.match(description:lower(), "undead") then
 --             family = "Undead";
 --         end
---         local criteriaString, criteriaType, completed, quantity, reqQuantity, charName, flags, assetID, quantityString, criteriaID = GetAchievementCriteriaInfo(achievementID, i);
+--         local criteriaString, criteriaType, completed, quantity, reqQuantity, charName, flags, assetID, quantityString, criteriaID = addon.GetAchievementCriteriaInfo(achievementID, i);
 --         if GetAchievementInfo(assetID) ~= nil then -- Means the assetID was not an achievementID but something else like a quest
 --             data.PrintCriteria(assetID, "A" .. tostring(id), i);
 --         else