+++ Plugins/ElvUI.lua
@@ -174,7 +174,7 @@ local function SkinAchievementsFrame(frame, engine, skins)
 		local textStrings, metas = 0, 0;
         local criteria, object;
 		for i = 1, numCriteria do
-			local _, criteriaType, completed, _, _, _, _, assetID = GetAchievementCriteriaInfo(id, i);
+			local _, criteriaType, completed, _, _, _, _, assetID = addon.GetAchievementCriteriaInfo(id, i);
 			if assetID and criteriaType == _G.CRITERIA_TYPE_ACHIEVEMENT then
 				metas = metas + 1;
 				criteria, object = self:GetMeta(metas), 'Label';